requirements.txt
asgiref==3.9.2
Django==5.2.6
sqlparse==0.5.3
tzdata==2025.2
djangorestframework==3.14.0
